Refactor models and add validation
