+++ app/models/offer.model.js
@@ -1,0 +1,11 @@
+var Schema = require('mongoose').Schema;
+
+var Offer = new Schema({
+    listingID: String,
+    buyerID: String,
+    sellerID: String,
+    ISBN: String,
+    date: Date,
+});
+
+module.exports = Offer;
