+++ app/models/report.model.js
@@ -1,0 +1,10 @@
+var Schema = require('mongoose').Schema;
+
+var Report = new Schema({
+    userID: { type: String, required: true },
+    reporterID: { type: String, required: true },
+    description: { type: String, required: true },
+    created: { type: Date, default: new Date() },
+});
+
+module.exports = Report;
