+++ app/models/school.model.js
@@ -1,0 +1,7 @@
+var Schema = require('mongoose').Schema;
+
+var School = new Schema({
+    // Name, logo?
+});
+
+module.exports = School;
